requirements.txt
discord
aiohttp==3.6.2
asgiref==3.2.10
async-timeout==3.0.1
asyncio==3.4.3
attrs==19.3.0
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.1
chardet==3.0.4
git+https://github.com/Rapptz/discord-ext-menus@84caae8038d0d3adc860957ccef05baeec2e2dd8
discord.py==1.4.1
Django==3.1
google-auth==1.21.0
google-auth-oauthlib==0.4.1
gspread==3.6.0
httplib2==0.18.1
idna==2.10
image==1.5.32
irc==19.0.0
jaraco.classes==3.1.0
jaraco.collections==3.0.0
jaraco.functools==3.0.1
jaraco.logging==3.0.0
jaraco.stream==3.0.0
jaraco.text==3.2.0
lavalink==3.1.2
more-itertools==8.4.0
multidict==4.7.6
oauth2client==4.1.3
oauthlib==3.1.0
Pillow==7.2.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyNaCl==1.3.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-oauthlib==1.3.0
rsa==4.6
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
six==1.15.0
sqlparse==0.3.1
tempora==4.0.0
twitchio==1.1.0
urllib3==1.25.10
vc==2018.7.10
websockets==8.1
wincertstore==0.2
yarl==1.5.1
youtube-dl==2020.7.28
